feat-added user object
Added user object for all the user details.
added logout functionality,
added authReducer in separate folder
