+++ src/frontend/reducer/index.js
@@ -3,3 +3,4 @@ export { filterReducer } from "./filterReducer/filterReducer";
 export { dataActions } from "./dataReducer/dataActionsConstant";
 export { dataReducer } from "./dataReducer/dataReducer";
 export { authActions } from "./authReducer/authActionsConstant";
+export { authReducer } from "./authReducer/authReducer";
